renamed bit_stuff
