+++ slsSupportLib/include/bit_utils.h
